move check chanel
